requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
supabase>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
scikit-learn>=1.3.0
numpy>=1.24.0
psycopg2-binary>=2.9.7
